pkg/ipmi * Move structures to structure.go * Move constants to constants.go
