+++ pkg/ipmi/ipmi.go
@@ -13,213 +13,24 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"log"
 	"math/rand"
-	"os"
-	"runtime"
 	"syscall"
 	"time"
 	"unsafe"
 
 	"github.com/vtolstov/go-ioctl"
-	"golang.org/x/sys/unix"
-)
-
-const (
-	_IPMI_BMC_CHANNEL                = 0xf
-	_IPMI_BUF_SIZE                   = 1024
-	_IPMI_IOC_MAGIC                  = 'i'
-	_IPMI_OPENIPMI_READ_TIMEOUT      = 15
-	_IPMI_SYSTEM_INTERFACE_ADDR_TYPE = 0x0c
-
-	// Net functions
-	_IPMI_NETFN_CHASSIS   NetFn = 0x0
-	_IPMI_NETFN_APP       NetFn = 0x6
-	_IPMI_NETFN_STORAGE   NetFn = 0xA
-	_IPMI_NETFN_TRANSPORT NetFn = 0xC
-
-	// IPM Device "Global" Commands
-	BMC_GET_DEVICE_ID Command = 0x01
-
-	// BMC Device and Messaging Commands
-	BMC_SET_WATCHDOG_TIMER     Command = 0x24
-	BMC_GET_WATCHDOG_TIMER     Command = 0x25
-	BMC_SET_GLOBAL_ENABLES     Command = 0x2E
-	BMC_GET_GLOBAL_ENABLES     Command = 0x2F
-	SET_SYSTEM_INFO_PARAMETERS Command = 0x58
-	BMC_ADD_SEL                Command = 0x44
-
-	// Chassis Device Commands
-	BMC_GET_CHASSIS_STATUS Command = 0x01
-
-	// SEL device Commands
-	BMC_GET_SEL_INFO Command = 0x40
-
-	//LAN Device Commands
-	BMC_GET_LAN_CONFIG Command = 0x02
-
-	IPM_WATCHDOG_NO_ACTION    = 0x00
-	IPM_WATCHDOG_SMS_OS       = 0x04
-	IPM_WATCHDOG_CLEAR_SMS_OS = 0x10
-
-	ADTL_SEL_DEVICE         = 0x04
-	EN_SYSTEM_EVENT_LOGGING = 0x08
-
-	// SEL
-	// STD_TYPE  = 0x02
-	OEM_NTS_TYPE = 0xFB
-
-	_SYSTEM_INFO_BLK_SZ = 16
-
-	_SYSTEM_FW_VERSION = 1
-
-	_ASCII = 0
-
-	// Set 62 Bytes (4 sets) as the maximal string length
-	strlenMax = 62
 )
 
 var (
-	_IPMICTL_RECEIVE_MSG_TRUNC = ioctl.IOWR(_IPMI_IOC_MAGIC, 11, uintptr(unsafe.Sizeof(recv{})))
-	_IPMICTL_SEND_COMMAND      = ioctl.IOR(_IPMI_IOC_MAGIC, 13, uintptr(unsafe.Sizeof(req{})))
+	_IPMICTL_RECEIVE_MSG_TRUNC = ioctl.IOWR(_IPMI_IOC_MAGIC, 11, uintptr(unsafe.Sizeof(response{})))
+	_IPMICTL_SEND_COMMAND      = ioctl.IOR(_IPMI_IOC_MAGIC, 13, uintptr(unsafe.Sizeof(request{})))
 
 	timeout = time.Second * 10
 )
 
 // IPMI represents access to the IPMI interface.
 type IPMI struct {
-	*os.File
-}
-
-// Command is the command code for a given message.
-type Command byte
-
-// NetFn is the network function of the class of message being sent.
-type NetFn byte
-
-// Msg is the full IPMI message to be sent.
-type Msg struct {
-	Netfn   NetFn
-	Cmd     Command
-	DataLen uint16
-	Data    unsafe.Pointer
-}
-
-type req struct {
-	addr    *systemInterfaceAddr
-	addrLen uint32
-	msgid   int64 //nolint:structcheck
-	msg     Msg
-}
-
-type recv struct {
-	recvType int32 //nolint:structcheck
-	addr     *systemInterfaceAddr
-	addrLen  uint32
-	msgid    int64 //nolint:structcheck
-	msg      Msg
-}
-
-type systemInterfaceAddr struct {
-	addrType int32
-	channel  int16
-	lun      byte //nolint:unused
-}
-
-// StandardEvent is a standard systemevent.
-//
-// The data in this event should follow IPMI spec
-type StandardEvent struct {
-	Timestamp    uint32
-	GenID        uint16
-	EvMRev       uint8
-	SensorType   uint8
-	SensorNum    uint8
-	EventTypeDir uint8
-	EventData    [3]uint8
-}
-
-// OEMTsEvent is a timestamped OEM-custom event.
-//
-// It holds 6 bytes of OEM-defined arbitrary data.
-type OEMTsEvent struct {
-	Timestamp        uint32
-	ManfID           [3]uint8
-	OEMTsDefinedData [6]uint8
-}
-
-// OEMNonTsEvent is a non-timestamped OEM-custom event.
-//
-// It holds 13 bytes of OEM-defined arbitrary data.
-type OEMNontsEvent struct {
-	OEMNontsDefinedData [13]uint8
-}
-
-// Event is included three kinds of events, Standard, OEM timestamped and OEM non-timestamped
-//
-// The record type decides which event should be used
-type Event struct {
-	RecordID   uint16
-	RecordType uint8
-	StandardEvent
-	OEMTsEvent
-	OEMNontsEvent
-}
-
-type setSystemInfoReq struct {
-	paramSelector byte
-	setSelector   byte
-	strData       [_SYSTEM_INFO_BLK_SZ]byte
-}
-
-type DevID struct {
-	DeviceID          byte
-	DeviceRevision    byte
-	FwRev1            byte
-	FwRev2            byte
-	IpmiVersion       byte
-	AdtlDeviceSupport byte
-	ManufacturerID    [3]byte
-	ProductID         [2]byte
-	AuxFwRev          [4]byte
-}
-
-type ChassisStatus struct {
-	CurrentPowerState byte
-	LastPowerEvent    byte
-	MiscChassisState  byte
-	FrontPanelButton  byte
-}
-
-type SELInfo struct {
-	Version     byte
-	Entries     uint16
-	FreeSpace   uint16
-	LastAddTime uint32
-	LastDelTime uint32
-	OpSupport   byte
-}
-
-func fdSet(fd uintptr, p *syscall.FdSet) {
-	p.Bits[fd/64] |= 1 << (uint(fd) % 64)
-}
-
-func ioctlSetReq(fd, name uintptr, req *req) error {
-	_, _, err := unix.Syscall(unix.SYS_IOCTL, fd, name, uintptr(unsafe.Pointer(req)))
-	runtime.KeepAlive(req)
-	if err != 0 {
-		return err
-	}
-	return nil
-}
-
-func ioctlGetRecv(fd, name uintptr, recv *recv) error {
-	_, _, err := unix.Syscall(unix.SYS_IOCTL, fd, name, uintptr(unsafe.Pointer(recv)))
-	runtime.KeepAlive(recv)
-	if err != 0 {
-		return err
-	}
-	return nil
+	DevAPI
 }
 
 // SendRecv sends the IPMI message, receives the response, and returns the
@@ -246,7 +57,7 @@ func (i *IPMI) RawSendRecv(msg Msg) ([]byte, error) {
 		addrType: _IPMI_SYSTEM_INTERFACE_ADDR_TYPE,
 		channel:  _IPMI_BMC_CHANNEL,
 	}
-	req := &req{
+	req := &request{
 		addr:    addr,
 		addrLen: uint32(unsafe.Sizeof(addr)),
 		msgid:   rand.Int63(),
@@ -255,7 +66,7 @@ func (i *IPMI) RawSendRecv(msg Msg) ([]byte, error) {
 
 	// Send request.
 	for {
-		switch err := ioctlSetReq(i.Fd(), _IPMICTL_SEND_COMMAND, req); {
+		switch err := i.DevAPI.SendRequest(req); {
 		case err == syscall.EINTR:
 			continue
 		case err != nil:
@@ -269,51 +80,21 @@ func (i *IPMI) RawSendRecv(msg Msg) ([]byte, error) {
 		Data:    unsafe.Pointer(&buf[0]),
 		DataLen: _IPMI_BUF_SIZE,
 	}
-	recv := &recv{
+	recv := &response{
 		addr:    req.addr,
 		addrLen: req.addrLen,
 		msg:     recvMsg,
 	}
 
+	data, err := i.DevAPI.ReceiveResponse(req.msgid, recv, buf)
+	if err != nil {
+		return nil, err
+	}
+	return data, nil
+
 	// This will be passed to the RawConn Read function. Until readMsg returns
 	// true, Read blocks and retries until the connection is ready for reading.
-	var result []byte
-	var rerr error
-	readMsg := func(fd uintptr) bool {
-		if err := ioctlGetRecv(fd, _IPMICTL_RECEIVE_MSG_TRUNC, recv); err != nil {
-			rerr = fmt.Errorf("ioctlGetRecv failed with %v", err)
-			return false
-		}
 
-		if recv.msgid != req.msgid {
-			log.Printf("Received wrong message. Trying again.")
-			return false
-		}
-
-		if recv.msg.DataLen >= _IPMI_BUF_SIZE {
-			rerr = fmt.Errorf("data length received too large: %d > %d", recv.msg.DataLen, _IPMI_BUF_SIZE)
-		} else if buf[0] != 0 {
-			rerr = fmt.Errorf("invalid response, expected first byte of response to be 0, got: %v", buf[0])
-		} else {
-			result = buf[:recv.msg.DataLen:recv.msg.DataLen]
-			rerr = nil
-		}
-		return true
-	}
-
-	// Read response.
-	conn, err := i.File.SyscallConn()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get file rawconn: %v", err)
-	}
-	if err := i.File.SetReadDeadline(time.Now().Add(timeout)); err != nil {
-		return nil, fmt.Errorf("failed to set read deadline: %v", err)
-	}
-	if err := conn.Read(readMsg); err != nil {
-		return nil, fmt.Errorf("failed to read rawconn: %v", err)
-	}
-
-	return result, rerr
 }
 
 func (i *IPMI) WatchdogRunning() (bool, error) {
@@ -552,4 +333,8 @@ func (i *IPMI) RawCmd(param []byte) ([]byte, error) {
 	msg.DataLen = uint16(len(param) - 2)
 
 	return i.RawSendRecv(msg)
+}
+
+func (i *IPMI) Close() error {
+	return i.DevAPI.GetFile().Close()
 }
